pyproject.toml
[tool.poetry]
name = "home-monitoring"
version = "0.1.0"
description = "Home sensors logging"
authors = ["tdurouchoux <thibault.durouchoux@gmail.com>"]
readme = "README.md"
packages = [{ include = "home_monitoring" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
influxdb-client = "^1.34.0"
reactivex = "^4.0.4"
pyyaml = "^6.0"
click = "^8.1.3"
tqdm = "^4.65.0"
omegaconf = "^2.3.0"
riche = "ˆ13.7.1"

[tool.poetry.group.sensorhub]

[tool.poetry.group.sensorhub.dependencies]
smbus = "^1.1.post2"

[tool.poetry.group.openweather]

[tool.poetry.group.openweather.dependencies]
requests = "^2.28.1"

[tool.poetry.group.teleinfo]

[tool.poetry.group.teleinfo.dependencies]
pyserial = "^3.5"

[tool.poetry.group.system_usage]

[tool.poetry.group.system_usage.dependencies]
psutil = "^5.9.4"

[tool.poetry.group.bme_sensor]

[tool.poetry.group.bme_sensor.dependencies]
pimoroni-bme280 = "^0.1.1"
bme680 = "^1.1.1"


[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple"
priority = 'default'

[[tool.poetry.source]]
name = "PyPI"
priority = 'primary'

[tool.poetry.scripts]
launch_monitoring = "home_monitoring.launch_monitoring:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
